requirements.txt
PyYAML>=6.0.1
aiogram>=3.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
